GitHubHealth/app/requirements.txt
altair==4.1.0
attrs==21.2.0
build==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
dominate==2.6.0
entrypoints==0.3
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.0
GitHubHealth==0.0.0.1637523240
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
numpy==1.21.4
packaging==20.9
pandas==1.3.4
pep517==0.12.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
toolz==0.11.2
uritemplate==3.0.1
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
zipp==3.6.0
